membenarkan float pembagian

+++ 1.cpp
@@ -26,10 +26,12 @@ int perkalian (int a,int b)
     return a * b;
 }
 
-int pembagian (int a,int b)
+float pembagian (int a,int b)
 {
     return a / b;
 }
+
+
